chore(protocol-designer): rename test-with-flow folder to __tests__
closes #4983
