+++ protocol-designer/src/step-generation/__tests__/consolidate.test.js
