+++ protocol-designer/src/step-generation/__tests__/dropAllTips.test.js
