+++ protocol-designer/src/step-generation/__tests__/engageMagnet.test.js
