+++ protocol-designer/src/step-generation/__tests__/fixtures/robotStateFixtures.js
