+++ protocol-designer/src/step-generation/__tests__/updateMagneticModule.test.js
